Move some test helpers from backend to dbapi
