+++ tests/test_lazy_connect/tests.py
@@ -4,9 +4,6 @@ from django.db import connections
 from salesforce.testrunner.example.models import User
 from requests.exceptions import ConnectionError
 
-# The test for `expectedFailure` decorator is put here, because it is not
-# nice to see '(expected failures=1)' in the main results
-from salesforce.backend.test_helpers import expectedFailureIf
 
 class LazyTest(TestCase):
     def test_lazy_connection(self):
@@ -25,13 +22,3 @@ class LazyTest(TestCase):
         # fix the host name and verify that the connection works now
         sf_conn.settings_dict.update(settings.ORIG_SALESFORCE_DB)
         self.assertGreater(len(users[:5]), 0)
-
-
-class TestExpectedFailure(TestCase):
-    @expectedFailureIf(False)
-    def test_condition_false(self):
-        assert True
-
-    @expectedFailureIf(True)
-    def test_condition_true(self):
-        assert False
